move reader toolbars into subdir
