+++ frontend/src/components/reader/toolbars/reader-book-change-nav-button.vue
