+++ frontend/src/components/reader/toolbars/reader-nav-button.vue
