+++ frontend/src/components/reader/toolbars/reader-title-toolbar.vue
